src/config/__init__.py
from .tools import SELECTED_SEARCH_ENGINE, SearchEngine
from .core import load_yaml_config

from dotenv import load_dotenv

# Load environment variables
load_dotenv()


__all__ = [
    # Other configurations
    "SELECTED_SEARCH_ENGINE",
    "SearchEngine",
    # Utilities
    "load_yaml_config",
]
