pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ptc-agent"
version = "0.1.0"
description = "ptc-agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    # LangChain core packages (v1)
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-text-splitters>=0.3.11",
    "langchain-community>=0.3.30",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain_google_genai>=2.0.0",
    "langchain-deepseek>=0.1.4",
    # LangGraph
    "langgraph>=0.3.5",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    # Other dependencies
    "readabilipy>=0.3.0",
    "python-dotenv>=1.0.1",
    "socksio>=1.0.0",
    "markdownify>=1.1.0",
    "crawl4ai>=0.4.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "websockets>=12.0",
    "sse-starlette>=2.1.0,<2.2.0",
    "pandas>=2.2.3",
    "numpy>=1.24,<3.0",
    "litellm>=1.63.11",
    "json-repair>=0.7.0",
    "jinja2>=3.1.3",
    "inquirerpy>=0.3.4",
    "textual>=0.47.0",
    "mcp>=1.6.0",
    "psycopg[binary,pool]>=3.2.9",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "python-json-logger>=2.0.7",
    "tushare==1.4.21",
    "yfinance==0.2.61",
    "matplotlib",
    "mplfinance==0.12.10b0",
    "tabulate",
    "alpha_vantage",
    "scikit-learn",
    "scipy",
    "statsmodels",
    "seaborn",
    "plotly",
    "arch",
    "pyportfolioopt",
    "QuantLib",
    "pymysql",
    "alibabacloud_oss_v2",
    "aiofiles==24.1.0",
    "querystring-parser==1.2.4",
    "lark==1.2.2",
    "ipykernel",
    "pandas_ta>=0.4.67b0",
    "setuptools",
    "rich>=14.2.0",
    "questionary>=2.0.0",
    "akshare>=1.13.0",
    "redis>=5.0.0",
    "langchain-qwq>=0.3.0",
    "edgartools>=5.7.2",
    "langsmith-fetch>=0.3.1",
    "structlog",
    "deepagents>=0.3.5",
    "daytona-sdk>=0.130.0",
    "boto3>=1.42.28",
    "tavily-python>=0.7.18",
    "ptc-cli",
]

[tool.uv.sources]
ptc-cli = { path = "libs/ptc-cli", editable = true }

[project.optional-dependencies]
dev = [
    "ruff",
    "black>=24.2.0",
    "langgraph-cli[inmem]>=0.2.10",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ptc_agent"]

[tool.uv]
required-version = ">=0.6.15"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
